Move spanUtilityMaps creation to a utility for testability
